Use regex to parse speedtest output
Using indices to parse strings is a bit fragile, so I thought I'd update it to use
regular expressions.  

Also, for users with slower internet service, pings of 100ms are not that uncommon.
I set `p` in that case to 0 thinking that it is more clear there was a connection issue
if ping/download/upload are all 0.