Address frequent best server failures on the hour
Speedtest frequently fails to find the best server on the
hour. Google searches revealed this to be due to high
server load. To address this, I added a random sleep (jitter),
and retry attempts. Both can be configured and disabled
via environment variables.

Additionally, I defaulted the results to 0.0 so you can see
in the graph when speedtest failures occurred.