Address various speedtest exceptions
Speedtest frequently fails to find the best server on the
hour. Google searches revealed this to be due to high
server load. To address this, I added a retry attempts
controlled by an environment variable.

I have also see speedtest fail to retrieve the config, so
I've included exception handling for it as well.

Additionally, I defaulted the results to 0.0 so you can see
in the graph when speedtest failures occurred.

Add missing utf-8 charset to index.html